// src/data/models/SustainabilityActivity.ts
import {
    Table,
    Column,
    Model,
    ForeignKey,
    DataType,
} from 'sequelize-typescript';
import User from './User';

@Table({ tableName: 'sustainabilityactivities', timestamps: true })
class SustainabilityActivity extends Model {
    @Column({
        type: DataType.INTEGER,
        autoIncrement: true,
        primaryKey: true,
    })
    public id!: number;

    @Column({
        type: DataType.INTEGER,
        allowNull: false,
    })
    public pointsAdded!: number;

    @Column({
        type: DataType.DECIMAL(10, 2),
        allowNull: false,
    })
    public carbonDioxideGenerated!: number;

    @Column({
        type: DataType.BOOLEAN,
        allowNull: false,
    })
    public isAutogenerated!: boolean;

    @ForeignKey(() => User)
    @Column({
        type: DataType.INTEGER,
        allowNull: false,
    })
    public userId!: number;

    // Relaciones entre entidades
    public user?: User;
}

export default SustainabilityActivity;